+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/PartialDataHandlerSpec.groovy
@@ -57,11 +57,11 @@ class PartialDataHandlerSpec extends Specification {
          * dim1 is missing four days of data internally, dim2 and dim3 are complete over the period and page_views is
          * starts inside the dim1 hole and goes to the end of the period.
          */
-        Map<Column, Set<Interval>> segmentIntervals = [
-                (new Column("userDeviceType")): buildIntervals(["2014-07-01/2014-07-09","2014-07-11/2014-07-29"]) as Set,
-                (new Column("property")): buildIntervals(["2014-07-01/2014-07-29"]) as Set,
-                (new Column("os")): buildIntervals(["2014-07-01/2014-07-29"]) as Set,
-                (new Column("page_views")): buildIntervals(["2014-07-04/2014-07-29"]) as Set
+        Map<String, Set<Interval>> segmentIntervals = [
+                'user_device_type': buildIntervals(["2014-07-01/2014-07-09","2014-07-11/2014-07-29"]) as Set,
+                'property': buildIntervals(["2014-07-01/2014-07-29"]) as Set,
+                'os': buildIntervals(["2014-07-01/2014-07-29"]) as Set,
+                'page_views': buildIntervals(["2014-07-04/2014-07-29"]) as Set
         ]
 
         table = new ConcretePhysicalTable(